refactor: use custom error handeling, validation of request body, id

+++ backend/src/controllers/ProductController.ts
@@ -1,5 +1,16 @@
-import { Response, Request } from "express";
+import { NextFunction, Response, Request } from "express";
+import mongoose from "mongoose";
 import { ProductService } from "../Service/ProductService";
+import {
+  BadRequestError,
+  ValidationError,
+  NotFoundError,
+} from "../utils/customErrors";
+import {
+  validateUpdateProduct,
+  validateAddProduct,
+} from "../validators/products";
+import { validateObjectId } from "../validators";
 
 export class ProductController {
   private productService: ProductService;
@@ -8,21 +19,31 @@ export class ProductController {
     this.productService = ProductService;
   }
 
-  async getProduct(req: Request, res: Response) {
+  async getProduct(req: Request, res: Response, next: NextFunction) {
     const productId = req.params._id;
+    if (!validateObjectId(productId)) {
+      return next(new BadRequestError("Invalid product ID format"));
+    }
     try {
-      const product = await this.productService.getProduct(productId);
-      res.status(200).json(product);
+      const productData = await this.productService.getProduct(productId);
+      if (!productData) {
+        return next(new NotFoundError("Product not found"));
+      }
+      res.status(200).json({
+        success: true,
+        message: "Product fetched successfully",
+        data: productData,
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
 
-  async getAllProducts(req: Request, res: Response) {
+  async getAllProducts(req: Request, res: Response, next: NextFunction) {
     try {
-      const { page, filter, category, limit, subCategory } = req.query;
-      const parsedPage = parseInt(page as string, 10);
-      const parsedLimit = parseInt(limit as string, 10);
+      const { page = 1, filter, category, limit = 9, subCategory } = req.query;
+      const parsedPage = Math.max(1, parseInt(page as string, 10));
+      const parsedLimit = Math.max(1, parseInt(limit as string, 10));
       const products = await this.productService.getAllProducts(
         parsedPage,
         parsedLimit,
@@ -30,54 +51,85 @@ export class ProductController {
         category as string,
         subCategory as string
       );
-      res.status(200).json(products);
+      res.status(200).json({
+        success: true,
+        message: "Products fetched successfully",
+        data: products,
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
 
-  async addProduct(req: Request, res: Response) {
+  async addProduct(req: Request, res: Response, next: NextFunction) {
     const product = req.body;
+    const { error } = validateAddProduct(product);
+    if (error) {
+      return next(new ValidationError(error.message));
+    }
     try {
       await this.productService.addProduct(product);
-      res.status(201).json({ message: "Product Added to DB!" });
+      res.status(201).json({
+        success: true,
+        message: "Product Added to DB!",
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
 
-  async deleteProduct(req: Request, res: Response) {
+  async deleteProduct(req: Request, res: Response, next: NextFunction) {
     const productId = req.params._id;
+    if (!validateObjectId(productId)) {
+      return next(new BadRequestError("Invalid product ID format"));
+    }
     try {
       await this.productService.deleteProduct(productId);
-      res.status(200).json({ message: "Product Deleted from DB!" });
+      res.status(200).json({
+        success: true,
+        message: "Product Deleted from DB!",
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
 
-  async updateProduct(req: Request, res: Response) {
+  async updateProduct(req: Request, res: Response, next: NextFunction) {
     const productId = req.params._id;
     const productData = req.body;
+    if (!validateObjectId(productId)) {
+      return next(new BadRequestError("Invalid product ID format"));
+    }
+    const { error } = validateUpdateProduct(productData);
+    if (error) {
+      return next(new ValidationError(error.message));
+    }
     try {
       await this.productService.updateProduct(productId, productData);
-      res.status(201).json({ message: "Product Updated!" });
+      res.status(201).json({
+        success: true,
+        message: "Product Updated!",
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
-  async getTopProducts(req: Request, res: Response) {
+  async getRandomProducts(req: Request, res: Response, next: NextFunction) {
     const { page = 1, limit = 3 } = req.query;
     const parsedPage = parseInt(page as string, 10);
     const parsedLimit = parseInt(limit as string, 10);
     try {
-      const result = await this.productService.getTopProducts(
+      const randomProducts = await this.productService.getRandomProducts(
         parsedPage,
         parsedLimit
       );
-      res.status(200).json(result);
+      res.status(200).json({
+        success: true,
+        message: "Random products fetched successfully",
+        data: randomProducts,
+      });
     } catch (error) {
-      res.status(400).json((error as Error).message);
+      next(error);
     }
   }
 }
